refactor: rename App.tsx to app.tsx
